requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.8
django-celery==3.1.17
django-uuidfield==0.5.0
djangorestframework==3.4.0
drf-nested-routers==0.11.1
drfdocs==0.0.11
httpie==0.9.4
kombu==3.0.35
Pygments==2.1.3
pytz==2016.6.1
redis==2.10.5
requests==2.10.0
